+++ src/parser/enrichers/spell/BallisticSmite.mjs
@@ -1,0 +1,10 @@
+/* eslint-disable class-methods-use-this */
+import DDBEnricherData from "../data/DDBEnricherData.mjs";
+
+export default class BallisticSmite extends DDBEnricherData {
+
+  get combineDamageTypes() {
+    return true;
+  }
+
+}