+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Next Up
 
+- Ballistic Smite parsing fixes
+
 # 6.5.2
 
 - Some monster weapons would import as features, such as the Light Hammer on the Bugbear Warrior, resulting in incorrect Melee/Range/Reach. @tminz